revamped the tests
